Fix PPU state machine
The get_mode function for checking the PPU mode was checking the lower 2
bits of LCDC, when it should have been checking STAT. This was making
putting it in a mode based off of basically a garbage value, which
completely broke the state machine logic.

Additionally, the state transitions were happening at the wrong points
in the scanline. They were checking if the position in the scanline was
equal to the duration of the mode, but they should really be comparing
against the absolute end-position of the mode in the scanline.