Implement DMA write lock and timer
The DMA transfer is *technically* ending at the right time, based on the
mooneye dma_timing test. However, I was only able to achieve this by
adding 2 m-cycles to its duration, so there's clearly a bug in the
timing logic somewhere. It's not worth fixing it in this state anyway, I
should probably get it to start transferring a byte at a time like it's
supposed to, and worry about it then.