Cycle-by-cycle prefixed execution
All of the prefixed instructions are now passing Blargg's mem timing
tests. The only issue is, it seems that in order to pass, I need to set
the m-cycle timings of reads and writes one m-cycle earlier than they
actually should be. there may be some sort of systematic error in the
prefixed execution timing that is causing this.