Implement cycle accurate for LD [HL] ops
I'm pretty much going all-in with making the instructions m-cycle
accurate at this point. The timing tests are never going to line up
if I don't start doing this across the board, because these very same
operations are being used to measure the timings. If the measuring tools
are off, the measurements will be as well.