Make progress on DMG acid
The right eye is rendering, but the part of the face underneath gets cut
off - when the stat register is not being updated.
When the stat register IS being updated, DMG acid is sending signals to
change wx, which toggles the rendering mode back from window to
background mid-frame.
However, with my current STAT interrupt implementation, updating every
single PPU tick, it's not working quite right. I am seeing part of the
background rendering return after the right eye is rendered, but the
entire face is still glitchy. So the STAT register and interrupt timings
need to be adjusted for sure. I already knew that they were off.